Add history to RingOther
